.test-infra/jenkins/committers_list_generator/requirements.txt
python-ldap
jinja2